+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 * Compatibility: Thread-safe patch for Qing's More Traits 2.
 * Compatibility: Thread-safe patch for DigitalStorage.
 * Compatibility: Thread-safe patch for MinifyEverything.
+* Compatibility: Thread-safe patch for Vivi Race / RPE Framework.
 
 ## 1.0.14  (2026-05-08)